+++ python/cudf/tests/test_dataframe.py
@@ -1086,6 +1086,24 @@ def test_from_pandas():
     pd.testing.assert_series_equal(s, gs.to_pandas())
 
 
+@pytest.mark.parametrize('dtypes', [int, float])
+def test_from_records(dtypes):
+    h_ary = np.ndarray(shape=(10, 4), dtype=dtypes)
+    rec_ary = h_ary.view(np.recarray)
+
+    gdf = gd.DataFrame.from_records(rec_ary, columns=['a', 'b', 'c', 'd'])
+    df = pd.DataFrame.from_records(rec_ary, columns=['a', 'b', 'c', 'd'])
+    assert isinstance(gdf, gd.DataFrame)
+
+    pd.testing.assert_frame_equal(df, gdf.to_pandas())
+
+    gdf = gd.DataFrame.from_records(rec_ary)
+    df = pd.DataFrame.from_records(rec_ary)
+    assert isinstance(gdf, gd.DataFrame)
+
+    pd.testing.assert_frame_equal(df, gdf.to_pandas())
+
+
 def test_from_gpu_matrix():
     h_ary = np.array([[1, 2, 3], [4, 5, 6]], np.int32)
     d_ary = rmm.to_device(h_ary)
@@ -1493,6 +1511,7 @@ def test_binops_series(pdf, gdf, binop):
 @pytest.mark.parametrize('func', [
     lambda df: df.empty,
     lambda df: df.x.empty,
+    lambda df: df.x.fillna(123, limit=None, method=None, axis=None),
 ])
 def test_unary_operators(func, pdf, gdf):
     p = func(pdf)
@@ -1954,7 +1973,7 @@ def test_dataframe_empty_sort_index():
     assert_eq(expect, got)
 
 
-@pytest.mark.parametrize('dtype', ['int8', 'int16', 'int32', 'int64',
+@pytest.mark.parametrize('dtype', ['bool', 'int8', 'int16', 'int32', 'int64',
                                    'float32', 'float64', 'datetime64[ms]',
                                    'category'])
 def test_dataframe_0_row_dtype(dtype):